refactor: refactor the constants
